# architecture/OpenApi/api.yaml
openapi: 3.0.0
info:
  description: >-
    This is the API for a course project at ITMO University. The theme of the
    course is "The goblin slayer contract system". Server on github.
  version: "1.1.0"
  title: Goblin Slayer Server API
servers:
  - description: Server
    url: 'http://paulrozhkin.ru:3464/'
...
paths:
...
  /api/v1/contracts/:
    #Получение списка всех контрактов
    get:
      tags:
        - contracts
      operationId: getContracts
      summary: Getting a list of contracts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: nameContract
          schema:
            type: string
          required: false
          description: Filtration by name
        - in: query
          name: customer
          schema:
            type: integer
          required: false
          description: Filtration by customer
        - in: query
          name: executor
          schema:
            type: integer
          required: false
          description: Filtration by executor
        - in: query
          name: rank
          schema:
            $ref: '#/components/schemas/ranksModel'
          required: false
          description: Filtration by rank. Returned all contracts that are equal to the required rank
        - in: query
          name: minRank
          schema:
            $ref: '#/components/schemas/ranksModel'
          required: false
          description: Filtration by minimum rank. Returned all contracts that are less than or equal to the required minimum rank. If a rank filter is set, minRank is ignored.
        - in: query
          name: contractStatus
          schema:
            $ref: '#/components/schemas/contractStatus'
          required: false
          description: Filtration by contractStatus.
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: Pagination parameter
        - in: query
          name: size
          schema:
            type: integer
          required: false
          description: Pagination parameter
      responses:
        '200':
          description: Getting a list of contracts
          content:
            application-json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/paginationInfo'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/contract'

        '400':
          description: Bad Reaquest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'

    #Создание новой модели контракта
    post:
      tags:
        - contracts
      operationId: createContract
      summary: Create the contract's model
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contractCreateModel'
        required: true
      responses:
        '200':
          description: The contract's model was created successfully.
          content:
            application-json:
              schema:
                $ref: '#/components/schemas/contract'
        '400':
          description: Bad Reaquest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /api/v1/contracts/{id}:
    #Получение модели контракта по ID
    get:
      tags:
        - contracts
      operationId: getСontractById
      summary: Getting the contract's model by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          description: 'Contract ID.'
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Contract model
          content:
            application-json:
              schema:
                $ref: '#/components/schemas/contract'
        '400':
          description: Bad Reaquest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'

      #Удаление модели пользователя по ID

    #Обновление модели контракта по ID
    put:
      tags:
        - contracts
      operationId: updateContractById
      summary: Update the contract's model
      security:
        - bearerAuth: []
      parameters:
        - name: id
          description: 'Contract ID.'
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contractUpdateModel'
        required: true
      responses:
        '200':
          description: The cotract's model was updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contract'
        '400':
          description: Bad Reaquest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /api/v1/contracts/{id}/perform/:
    # Начать исполнение контракта авантюристом
    post:
      tags:
        - contracts
      operationId: startPerformContractByAdventurer
      summary: Start performing contact
      security:
        - bearerAuth: []
      parameters:
        - name: id
          description: 'Contract ID.'
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The adventurer model was updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contract'
        '400':
          description: Bad Reaquest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '403':
          description: "Forbidden. Only adventurer can start perform contact."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /api/v1/contracts/{id}/performed/:
    # Завершить исполнение контракта авантюристом
    post:
      tags:
        - contracts
      operationId: performedContractByAdventurer
      summary: Stop performing contract if completed
      security:
        - bearerAuth: []
      parameters:
        - name: id
          description: 'Contract ID.'
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contractPerformedModel'
        required: true
      responses:
        '200':
          description: The adventurer model was updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contract'
        '400':
          description: Bad Reaquest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '403':
          description: "Forbidden. Only adventurer can stop performing contact."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /api/v1/contracts/{id}/cancel/:
    # Отменить исполнение контракта текущим авантюристом
    post:
      tags:
        - contracts
      operationId: cancelContractByAdventurer
      summary: Cancel contact
      security:
        - bearerAuth: []
      parameters:
        - name: id
          description: 'Contract ID.'
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contractCancellationModel'
        required: true
      responses:
        '200':
          description: The contract was cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contract'
        '400':
          description: Bad Reaquest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '403':
          description: "Forbidden. Only adventurer, registrar and admin can cancel contact."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /api/v1/files/:
    post:
      tags:
        - files
      operationId: uploadFile
      summary: 'Uploads a file.'
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: # Part 2 - file
                  type: string
                  format: binary
      responses:
        '200':
          description: File uploaded
          content:
            application-json:
              schema:
                $ref: '#/components/schemas/fileModel'
        '400':
          description: Bad Reaquest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /api/v1/files/{id}:
    #Получение файла по id
    get:
      tags:
        - files
      operationId: downloadFileById
      summary: 'Download file'
      parameters:
        - name: id
          description: 'File id.'
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A file
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
components:
  schemas:
...
    paginationInfo:
      type: object
      properties:
        currentPage:
          description: 'Pagination current page'
          type: integer
          example: '0'
        totalItems:
          description: 'Pagination total items'
          type: integer
          format: int64
          example: '45'
        totalPages:
          description: 'Pagination total pages'
          type: integer
          example: '15'
...
    contractStatus:
      description: 'Contract status in the system.'
      type: string
      enum:
        - Created
        - Rejected
        - Accepted
        - Performing
        - Performed
        - Completed
      example:
        Created
    ranksModel:
      description: 'The rank of adventurer.'
      type: string
      enum:
        - Platinum
        - Gold
        - Silver
        - Bronze
        - Ruby
        - Emerald
        - Sapphire
        - Steel
        - Obsidian
        - Porcelain
      example:
        Emerald
    contract:
      allOf:
        - $ref: '#/components/schemas/contractIdModel'
        - $ref: '#/components/schemas/contractModel'
    contractIdModel:
      type: object
      properties:
        id:
          description: 'Unique identifier of the user.'
          type: integer
          example: 15
    contractCreateModel:
      type: object
      properties:
        customer:
          description: 'Contract customer.'
          type: integer
          example: 12
        nameContract:
          description: 'Name of contract.'
          type: string
          example: 'Killing goblins in the mountains'
        reward:
          description: 'Reward for completed task.'
          type: integer
          example: 500
        address:
          description: 'The address at which the contract is to be performed.'
          type: boolean
          example: 'Novoizmaylovsky Prospekt, 21k4, Saint Petersburg, Russia'
        description:
          description: 'Description of the contract submitted by the contractor.'
          type: string
          example: 'You need to kill 10 goblins and 1 Hobbs in the winter cave. They are poorly prepared, do not have armor and weapons.'
        requestComment:
          description: 'Review of the Guild Registrar on the submitted contract.'
          type: string
          example: 'The abilities of goblins and Hobbs are not described.'
        icon:
          description: 'Main icon id of contract'
          type: integer
          example: '3'
    contractUpdateModel:
      allOf:
        - type: object
          properties:
            executor:
              description: 'Contract executor.'
              type: integer
              example: 25
            reward:
              description: 'Reward for completed task.'
              type: integer
              example: 500
            minRank:
              $ref: '#/components/schemas/ranksModel'
            address:
              description: 'The address at which the contract is to be performed.'
              type: string
              example: 'Novoizmaylovsky Prospekt, 21k4, Saint Petersburg, Russia'
            contractStatus:
              $ref: '#/components/schemas/contractStatus'
            description:
              description: 'Description of the contract submitted by the contractor.'
              type: string
              example: 'You need to kill 10 goblins and 1 Hobbs in the winter cave. They are poorly prepared, do not have armor and weapons.'
            requestComment:
              description: 'Review of the Guild Registrar on the submitted contract.'
              type: string
              example: 'The abilities of goblins and Hobbs are not described.'
            registrarComment:
              description: 'Review of the Registrar.'
              type: string
              example: 'The abilities of goblins and Hobbs are not described.'
            icon:
              description: 'Main icon id of contract'
              type: integer
              example: '3'
    contractModel:
      allOf:
        - type: object
          properties:
            customer:
              description: 'Contract customer.'
              type: integer
              example: 12
            executor:
              description: 'Contract executor.'
              type: integer
              example: 25
            nameContract:
              description: 'Name of contract.'
              type: string
              example: 'Killing goblins in the mountains'
            reward:
              description: 'Reward for completed task.'
              type: integer
              example: 500
            minRank:
              $ref: '#/components/schemas/ranksModel'
            address:
              description: 'The address at which the contract is to be performed.'
              type: string
              example: 'Novoizmaylovsky Prospekt, 21k4, Saint Petersburg, Russia'
            createTime:
              description: 'Date the contract was created.'
              type: string
              format: date-time
              example: '2017-07-21T17:32:28Z'
            contractStatus:
              $ref: '#/components/schemas/contractStatus'
            description:
              description: 'Description of the contract submitted by the contractor.'
              type: string
              example: 'You need to kill 10 goblins and 1 Hobbs in the winter cave. They are poorly prepared, do not have armor and weapons.'
            requestComment:
              description: 'Review of the Guild Registrar on the submitted contract.'
              type: string
              example: 'The abilities of goblins and Hobbs are not described.'
            registrarComment:
              description: 'Review of the Registrar.'
              type: string
              example: 'The abilities of goblins and Hobbs are not described.'
            performedComment:
              description: 'Revocation of the adventurer is contract performance.'
              type: string
              example: 'All dead.'
            cancellationComment:
              description: 'Cancellation comment.'
              type: string
              example: 'The contract was too difficult.'
            icon:
              description: 'Main icon id of contract'
              type: integer
              example: '3'
...
    contractPerformedModel:
      type: object
      properties:
        performedComment:
          description: 'Unique identifier of the contractNotification.'
          type: string
          example: "All the goblins are killed."
    contractCancellationModel:
      type: object
      properties:
        cancellationComment:
          description: 'Unique identifier of the contractNotification.'
          type: string
          example: "The contract was too difficult."
...
    fileModel:
      type: object
      properties:
        id:
          type: integer
          description: 'File id.'
        fileName:
          type: string
          description: 'File name.'
        createTime:
          description: 'Date the file was created.'
          type: string
          format: date-time
          example: '2017-07-21T17:32:28Z'
    errorModel:
      type: object
      properties:
        message:
          type: string
          description: 'Error description.'
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes